Cudatoolkit build recipe, update and automate 7.5 and 8.0.
This code builds the "cudatoolkit" packages. A condarecipe exists
for each of 7.5 and 8.0, this is just needed to separate the
incremental build numbering, the actual code used to build the
packages is the same.

This new code should enable a more configurable CUDA toolkit
package build. As a result of the configuration capacity a yaml
file is now dumped into the library directory of the package that
contains the configuration dictionary such that there is
tracibility from the package to the binary source.

To add additional cuda versions, add to the dictionary defined at
the top of build.py.

This is not an ideal method of assembling the package, but is
sufficient for reproducability.

This is a first cut, patches welcome.