Accommodate NvToolsExt.
This adds the environment variable "NVTOOLSEXT_INSTALL_PATH" as
a pass through variable to the conda build scripts. This is to
accommodate the windows installers not having the nvtoolsext DLL
as an extractable item and allow assembly from an already present
installation, the location of the dll is not the install directory!